requirements.txt
fastapi==0.68.2
httpx==0.23.0
pydantic==1.9.0
uvicorn==0.15.0
whisper==1.0.0
ffmpeg-python==0.2.0
openai==1.6.1